+++ bonus/sources/ft_colouring_2_bonus.c
@@ -1,22 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_colouring_2.c                                   :+:      :+:    :+:   */
+/*   ft_colouring_2_bonus.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/22 18:00:07 by smagdela          #+#    #+#             */
-/*   Updated: 2021/12/23 11:52:05 by smagdela         ###   ########.fr       */
+/*   Updated: 2021/12/23 13:42:40 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "fractol.h"
+#include "fractol_bonus.h"
 
 int	color_grayscale(int n, int max_iter)
 {
 	int	x;
 
-	x = -255 * ((float)n / (float)max_iter) + 255;
+	x = -255 * ((double)n / (double)max_iter) + 255;
 	return (x * 0x010000 + x * 0x000100 + x);
 }
 
@@ -25,12 +25,12 @@ int	color_multi(int n, int max_iter)
 	int		r;
 	int		g;
 	int		b;
-	float	x;
+	double	x;
 
 	r = 0;
 	g = 0;
 	b = 0;
-	x = (float)n / (float)max_iter;
+	x = (double)n / (double)max_iter;
 	if (x >= 0 && x <= 0.5)
 	{
 		r = -510 * x - 255;
@@ -49,10 +49,10 @@ int	color_lagoon(int n, int max_iter)
 	int		r;
 	int		g;
 	int		b;
-	float	x;
+	double	x;
 
 	r = 0;
-	x = (float)n / (float)max_iter;
+	x = (double)n / (double)max_iter;
 	b = 127.5 * x + 127.5;
 	g = 255 * x;
 	if (x >= 0.9)
@@ -65,11 +65,11 @@ int	color_flame(int n, int max_iter)
 	int		r;
 	int		g;
 	int		b;
-	float	x;
+	double	x;
 
 	r = 255;
 	b = 0;
-	x = (float)n / (float)max_iter;
+	x = (double)n / (double)max_iter;
 	g = 510 * x;
 	if (x >= 0.5)
 		g = -510 * x + 510;
